+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Changes to Calva.
 
 ## [Unreleased]
 - Fix: [Allow keybindings to target when the cursor is inside a comment or a string](https://github.com/BetterThanTomorrow/calva/issues/1023)
+- Fix: [Navigating to a definition in a jar file throws error in console](https://github.com/BetterThanTomorrow/calva/issues/1047)
+
+## [2.0.176] - 2021-02-24
+- Revert switch to cljs for lsp, until [the issue with released cljs/js interop](https://github.com/BetterThanTomorrow/calva/issues/1044) has been fixed
+
+## [2.0.174] - 2021-02-24
+- [Translate clojure-lsp integration to cljs](https://github.com/BetterThanTomorrow/calva/issues/1025)
+
+## [2.0.173] - 2021-02-21
+- Fix [Connect ”not in project” glitches](https://github.com/BetterThanTomorrow/calva/issues/814)
+- [Add a ”Start Standalone REPL” commands](https://github.com/BetterThanTomorrow/calva/issues/1003)
+- [Add a configuration option to disable diagnostics](https://github.com/BetterThanTomorrow/calva/pull/981)
 
 ## [2.0.171] - 2021-02-10
 - Update clojure-lsp to version 2021.02.09-18.28.06 (Fix: [Auto completion does not work in clojure-lsp only mode (no repl connection)](https://github.com/BetterThanTomorrow/calva/issues/996#issuecomment-776148282))